+++ src/components/modules/blockManager.ts
@@ -11,7 +11,8 @@ import Module from '../__module';
 import $ from '../dom';
 import * as _ from '../utils';
 import Blocks from '../blocks';
-import { BlockToolConstructable, BlockToolData, PasteEvent } from '../../../types';
+import { BlockToolData, PasteEvent } from '../../../types';
+import { BlockTuneData } from '../../../types/block-tunes/block-tune-data';
 
 /**
  * @typedef {BlockManager} BlockManager
@@ -184,7 +185,7 @@ export default class BlockManager extends Module {
     });
 
     /** Copy event */
-    this.Editor.Listeners.on(
+    this.listeners.on(
       document,
       'copy',
       (e: ClipboardEvent) => this.Editor.BlockEvents.handleCommandC(e)
@@ -215,21 +216,26 @@ export default class BlockManager extends Module {
    *
    * @param {object} options - block creation options
    * @param {string} options.tool - tools passed in editor config {@link EditorConfig#tools}
+   * @param {string} [options.id] - unique id for this block
    * @param {BlockToolData} [options.data] - constructor params
    *
    * @returns {Block}
    */
-  public composeBlock({ tool, data = {} }: {tool: string; data?: BlockToolData}): Block {
+  public composeBlock({
+    tool: name,
+    data = {},
+    id = undefined,
+    tunes: tunesData = {},
+  }: {tool: string; id?: string; data?: BlockToolData; tunes?: {[name: string]: BlockTuneData}}): Block {
     const readOnly = this.Editor.ReadOnly.isEnabled;
-    const settings = this.Editor.Tools.getToolSettings(tool);
-    const Tool = this.Editor.Tools.available[tool] as BlockToolConstructable;
+    const tool = this.Editor.Tools.blockTools.get(name);
     const block = new Block({
-      name: tool,
+      id,
       data,
-      Tool,
-      settings,
+      tool,
       api: this.Editor.API,
       readOnly,
+      tunesData,
     });
 
     if (!readOnly) {
@@ -243,26 +249,31 @@ export default class BlockManager extends Module {
    * Insert new block into _blocks
    *
    * @param {object} options - insert options
-   * @param {string} options.tool - plugin name, by default method inserts the default block type
-   * @param {object} options.data - plugin data
-   * @param {number} options.index - index where to insert new Block
-   * @param {boolean} options.needToFocus - flag shows if needed to update current Block index
-   * @param {boolean} options.replace - flag shows if block by passed index should be replaced with inserted one
+   * @param {string} [options.id] - block's unique id
+   * @param {string} [options.tool] - plugin name, by default method inserts the default block type
+   * @param {object} [options.data] - plugin data
+   * @param {number} [options.index] - index where to insert new Block
+   * @param {boolean} [options.needToFocus] - flag shows if needed to update current Block index
+   * @param {boolean} [options.replace] - flag shows if block by passed index should be replaced with inserted one
    *
    * @returns {Block}
    */
   public insert({
+    id = undefined,
     tool = this.config.defaultBlock,
     data = {},
     index,
     needToFocus = true,
     replace = false,
+    tunes = {},
   }: {
+    id?: string;
     tool?: string;
     data?: BlockToolData;
     index?: number;
     needToFocus?: boolean;
     replace?: boolean;
+    tunes?: {[name: string]: BlockTuneData};
   } = {}): Block {
     let newIndex = index;
 
@@ -271,8 +282,10 @@ export default class BlockManager extends Module {
     }
 
     const block = this.composeBlock({
+      id,
       tool,
       data,
+      tunes,
     });
 
     this._blocks.insert(newIndex, block, replace);
@@ -509,6 +522,17 @@ export default class BlockManager extends Module {
   }
 
   /**
+   * Returns the Block by passed id
+   *
+   * @param id - id of block to get
+   *
+   * @returns {Block}
+   */
+  public getBlockById(id): Block {
+    return this._blocks.array.find(block => block.id === id);
+  }
+
+  /**
    * Get Block instance by html element
    *
    * @param {Node} element - html element to get Block by
@@ -559,11 +583,10 @@ export default class BlockManager extends Module {
    * 1) Find first-level Block from passed child Node
    * 2) Mark it as current
    *
-   *  @param {Node} childNode - look ahead from this node.
-   *
-   *  @throws Error  - when passed Node is not included at the Block
+   * @param {Node} childNode - look ahead from this node.
+   * @returns {Block | undefined} can return undefined in case when the passed child note is not a part of the current editor instance
    */
-  public setCurrentBlockByChildNode(childNode: Node): Block {
+  public setCurrentBlockByChildNode(childNode: Node): Block | undefined {
     /**
      * If node is Text TextNode
      */
@@ -573,23 +596,36 @@ export default class BlockManager extends Module {
 
     const parentFirstLevelBlock = (childNode as HTMLElement).closest(`.${Block.CSS.wrapper}`);
 
-    if (parentFirstLevelBlock) {
-      /**
-       * Update current Block's index
-       *
-       * @type {number}
-       */
-      this.currentBlockIndex = this._blocks.nodes.indexOf(parentFirstLevelBlock as HTMLElement);
-
-      /**
-       * Update current block active input
-       */
-      this.currentBlock.updateCurrentInput();
-
-      return this.currentBlock;
-    } else {
-      throw new Error('Can not find a Block from this child Node');
+    if (!parentFirstLevelBlock) {
+      return;
     }
+
+    /**
+     * Support multiple Editor.js instances,
+     * by checking whether the found block belongs to the current instance
+     *
+     * @see {@link Ui#documentTouched}
+     */
+    const editorWrapper = parentFirstLevelBlock.closest(`.${this.Editor.UI.CSS.editorWrapper}`);
+    const isBlockBelongsToCurrentInstance = editorWrapper?.isEqualNode(this.Editor.UI.nodes.wrapper);
+
+    if (!isBlockBelongsToCurrentInstance) {
+      return;
+    }
+
+    /**
+     * Update current Block's index
+     *
+     * @type {number}
+     */
+    this.currentBlockIndex = this._blocks.nodes.indexOf(parentFirstLevelBlock as HTMLElement);
+
+    /**
+     * Update current block active input
+     */
+    this.currentBlock.updateCurrentInput();
+
+    return this.currentBlock;
   }
 
   /**
@@ -691,9 +727,7 @@ export default class BlockManager extends Module {
    */
   public async destroy(): Promise<void> {
     await Promise.all(this.blocks.map((block) => {
-      if (_.isFunction(block.tool.destroy)) {
-        return block.tool.destroy();
-      }
+      return block.destroy();
     }));
   }
 
@@ -707,7 +741,7 @@ export default class BlockManager extends Module {
 
     this.readOnlyMutableListeners.on(block.holder, 'keydown', (event: KeyboardEvent) => {
       BlockEvents.keydown(event);
-    }, true);
+    });
 
     this.readOnlyMutableListeners.on(block.holder, 'keyup', (event: KeyboardEvent) => {
       BlockEvents.keyup(event);