+++ src/components/modules/blockSelection.ts
@@ -9,9 +9,11 @@ import Module from '../__module';
 import Block from '../block';
 import * as _ from '../utils';
 import $ from '../dom';
+import Shortcuts from '../utils/shortcuts';
 
 import SelectionUtils from '../selection';
 import { SanitizerConfig } from '../../../types/configs';
+import { clean } from '../utils/sanitizer';
 
 /**
  *
@@ -145,8 +147,6 @@ export default class BlockSelection extends Module {
    * to select all and copy them
    */
   public prepare(): void {
-    const { Shortcuts } = this.Editor;
-
     this.selection = new SelectionUtils();
 
     /**
@@ -181,6 +181,7 @@ export default class BlockSelection extends Module {
 
         this.handleCommandA(event);
       },
+      on: this.Editor.UI.nodes.redactor,
     });
   }
 
@@ -285,7 +286,7 @@ export default class BlockSelection extends Module {
    *
    * @returns {Promise<void>}
    */
-  public async copySelectedBlocks(e: ClipboardEvent): Promise<void> {
+  public copySelectedBlocks(e: ClipboardEvent): Promise<void> {
     /**
      * Prevent default copy
      */
@@ -297,14 +298,12 @@ export default class BlockSelection extends Module {
       /**
        * Make <p> tag that holds clean HTML
        */
-      const cleanHTML = this.Editor.Sanitizer.clean(block.holder.innerHTML, this.sanitizerConfig);
+      const cleanHTML = clean(block.holder.innerHTML, this.sanitizerConfig);
       const fragment = $.make('p');
 
       fragment.innerHTML = cleanHTML;
       fakeClipboard.appendChild(fragment);
     });
-
-    const savedData = await Promise.all(this.selectedBlocks.map((block) => block.save()));
 
     const textPlain = Array.from(fakeClipboard.childNodes).map((node) => node.textContent)
       .join('\n\n');
@@ -312,7 +311,16 @@ export default class BlockSelection extends Module {
 
     e.clipboardData.setData('text/plain', textPlain);
     e.clipboardData.setData('text/html', textHTML);
-    e.clipboardData.setData(this.Editor.Paste.MIME_TYPE, JSON.stringify(savedData));
+
+    return Promise
+      .all(this.selectedBlocks.map((block) => block.save()))
+      .then(savedData => {
+        try {
+          e.clipboardData.setData(this.Editor.Paste.MIME_TYPE, JSON.stringify(savedData));
+        } catch (err) {
+          // In Firefox we can't set data in async function
+        }
+      });
   }
 
   /**
@@ -361,10 +369,8 @@ export default class BlockSelection extends Module {
    * De-registers Shortcut CMD+A
    */
   public destroy(): void {
-    const { Shortcuts } = this.Editor;
-
     /** Selection shortcut */
-    Shortcuts.remove('CMD+A');
+    Shortcuts.remove(this.Editor.UI.nodes.redactor, 'CMD+A');
   }
 
   /**