+++ src/components/modules/toolbar/blockSettings.ts
@@ -30,7 +30,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
    *
    * @returns {{opened: string, closed: string}}
    */
-  public get events(): {opened: string; closed: string} {
+  public get events(): { opened: string; closed: string } {
     return {
       opened: 'block-settings-opened',
       closed: 'block-settings-closed',
@@ -42,7 +42,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
    *
    * @returns {{wrapper, wrapperOpened, toolSettings, defaultSettings, button}}
    */
-  public get CSS(): {[name: string]: string} {
+  public get CSS(): { [name: string]: string } {
     return {
       // Settings Panel
       wrapper: 'ce-settings',
@@ -144,10 +144,10 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
     /**
      * Add default settings that presents for all Blocks
      */
-    this.addDefaultSettings();
+    this.addTunes();
 
     /** Tell to subscribers that block settings is opened */
-    this.Editor.Events.emit(this.events.opened);
+    this.eventsDispatcher.emit(this.events.opened);
 
     this.flipper.activate(this.blockTunesButtons);
   }
@@ -171,12 +171,19 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
 
     this.selection.clearSaved();
 
+    /**
+     * Remove highlighted content of a Block we are working with
+     */
+    if (!this.Editor.CrossBlockSelection.isCrossBlockSelectionStarted && this.Editor.BlockManager.currentBlock) {
+      this.Editor.BlockManager.currentBlock.selected = false;
+    }
+
     /** Clear settings */
     this.nodes.toolSettings.innerHTML = '';
     this.nodes.defaultSettings.innerHTML = '';
 
     /** Tell to subscribers that block settings is closed */
-    this.Editor.Events.emit(this.events.closed);
+    this.eventsDispatcher.emit(this.events.closed);
 
     /** Clear cached buttons */
     this.buttons = [];
@@ -222,16 +229,21 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
    * Add Tool's settings
    */
   private addToolSettings(): void {
-    if (_.isFunction(this.Editor.BlockManager.currentBlock.tool.renderSettings)) {
-      $.append(this.nodes.toolSettings, this.Editor.BlockManager.currentBlock.tool.renderSettings());
+    const settingsElement = this.Editor.BlockManager.currentBlock.renderSettings();
+
+    if (settingsElement) {
+      $.append(this.nodes.toolSettings, settingsElement);
     }
   }
 
   /**
-   * Add default settings
+   * Add tunes: provided by user and default ones
    */
-  private addDefaultSettings(): void {
-    $.append(this.nodes.defaultSettings, this.Editor.BlockManager.currentBlock.renderTunes());
+  private addTunes(): void {
+    const [toolTunes, defaultTunes] = this.Editor.BlockManager.currentBlock.renderTunes();
+
+    $.append(this.nodes.toolSettings, toolTunes);
+    $.append(this.nodes.defaultSettings, defaultTunes);
   }
 
   /**