+++ src/components/modules/toolbar/inline.ts
@@ -2,10 +2,17 @@ import Module from '../../__module';
 import $ from '../../dom';
 import SelectionUtils from '../../selection';
 import * as _ from '../../utils';
-import { InlineTool, InlineToolConstructable, ToolConstructable, ToolSettings } from '../../../../types';
+import { InlineTool as IInlineTool, EditorConfig } from '../../../../types';
 import Flipper from '../../flipper';
 import I18n from '../../i18n';
 import { I18nInternalNS } from '../../i18n/namespace-internal';
+import Shortcuts from '../../utils/shortcuts';
+import Tooltip from '../../utils/tooltip';
+import { ModuleConfig } from '../../../types-internal/module-config';
+import EventsDispatcher from '../../utils/events';
+import InlineTool from '../../tools/inline';
+import { CommonInternalSettings } from '../../tools/base';
+import BlockTool from '../../tools/block';
 
 /**
  * Inline Toolbar elements
@@ -64,9 +71,11 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
   private readonly toolbarVerticalMargin: number = 5;
 
   /**
+   * TODO: Get rid of this
+   *
    * Currently visible tools instances
    */
-  private toolsInstances: Map<string, InlineTool>;
+  private toolsInstances: Map<string, IInlineTool>;
 
   /**
    * Buttons List
@@ -86,6 +95,24 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    * Instance of class that responses for leafing buttons by arrows/tab
    */
   private flipper: Flipper = null;
+
+  /**
+   * Tooltip utility Instance
+   */
+  private tooltip: Tooltip;
+  /**
+   * @class
+   * @param {object} moduleConfiguration - Module Configuration
+   * @param {EditorConfig} moduleConfiguration.config - Editor's config
+   * @param {EventsDispatcher} moduleConfiguration.eventsDispatcher - Editor's event dispatcher
+   */
+  constructor({ config, eventsDispatcher }: ModuleConfig) {
+    super({
+      config,
+      eventsDispatcher,
+    });
+    this.tooltip = new Tooltip();
+  }
 
   /**
    * Toggles read-only mode
@@ -185,7 +212,13 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     }
 
     this.nodes.wrapper.classList.remove(this.CSS.inlineToolbarShowed);
-    this.toolsInstances.forEach((toolInstance) => {
+    Array.from(this.toolsInstances.entries()).forEach(([name, toolInstance]) => {
+      const shortcut = this.getToolShortcut(name);
+
+      if (shortcut) {
+        Shortcuts.remove(this.Editor.UI.nodes.redactor, shortcut);
+      }
+
       /**
        * @todo replace 'clear' with 'destroy'
        */
@@ -265,80 +298,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     }
 
     this.removeAllNodes();
-  }
-
-  /**
-   * Returns inline toolbar settings for a particular tool
-   *
-   * @param {string} toolName - user specified name of tool
-   * @returns {string[] | boolean} array of ordered tool names or false
-   */
-  private getInlineToolbarSettings(toolName): string[] | boolean {
-    const toolSettings = this.Editor.Tools.getToolSettings(toolName);
-
-    /**
-     * InlineToolbar property of a particular tool
-     */
-    const settingsForTool = toolSettings[this.Editor.Tools.USER_SETTINGS.ENABLED_INLINE_TOOLS];
-
-    /**
-     * Whether to enable IT for a particular tool is the decision of the editor user.
-     * He can enable it by the inlineToolbar settings for this tool. To enable, he should pass true or strings[]
-     */
-    const enabledForTool = settingsForTool === true || Array.isArray(settingsForTool);
-
-    /**
-     * Disabled by user
-     */
-    if (!enabledForTool) {
-      return false;
-    }
-
-    /**
-     * 1st priority.
-     *
-     * If user pass the list of inline tools for the particular tool, return it.
-     */
-    if (Array.isArray(settingsForTool)) {
-      return settingsForTool;
-    }
-
-    /**
-     * 2nd priority.
-     *
-     * If user pass just 'true' for tool, get common inlineToolbar settings
-     * - if common settings is an array, use it
-     * - if common settings is 'true' or not specified, get default order
-     */
-
-    /**
-     * Common inlineToolbar settings got from the root of EditorConfig
-     */
-    const commonInlineToolbarSettings = this.config.inlineToolbar;
-
-    /**
-     * If common settings is an array, use it
-     */
-    if (Array.isArray(commonInlineToolbarSettings)) {
-      return commonInlineToolbarSettings;
-    }
-
-    /**
-     * If common settings is 'true' or not specified (will be set as true at core.ts), get the default order
-     */
-    if (commonInlineToolbarSettings === true) {
-      const defaultToolsOrder: string[] = Object.entries(this.Editor.Tools.available)
-        .filter(([name, tool]) => {
-          return tool[this.Editor.Tools.INTERNAL_SETTINGS.IS_INLINE];
-        })
-        .map(([name, tool]) => {
-          return name;
-        });
-
-      return defaultToolsOrder;
-    }
-
-    return false;
+    this.tooltip.destroy();
   }
 
   /**
@@ -357,7 +317,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     this.nodes.actions = $.make('div', this.CSS.actionsWrapper);
 
     // To prevent reset of a selection when click on the wrapper
-    this.Editor.Listeners.on(this.nodes.wrapper, 'mousedown', (event) => {
+    this.listeners.on(this.nodes.wrapper, 'mousedown', (event) => {
       const isClickedOnActionsWrapper = (event.target as Element).closest(`.${this.CSS.actionsWrapper}`);
 
       // If click is on actions wrapper,
@@ -449,12 +409,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       return false;
     }
 
-    /**
-     * getInlineToolbarSettings could return an string[] (order of tools) or false (Inline Toolbar disabled).
-     */
-    const inlineToolbarSettings = this.getInlineToolbarSettings(currentBlock.name);
-
-    return inlineToolbarSettings !== false;
+    return currentBlock.tool.inlineTools.size !== 0;
   }
 
   /**
@@ -479,7 +434,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
 
     this.nodes.togglerAndButtonsWrapper.appendChild(this.nodes.conversionToggler);
 
-    this.Editor.Listeners.on(this.nodes.conversionToggler, 'click', () => {
+    this.listeners.on(this.nodes.conversionToggler, 'click', () => {
       this.Editor.ConversionToolbar.toggle((conversionToolbarOpened) => {
         /**
          * When ConversionToolbar is opening on activated InlineToolbar flipper
@@ -498,7 +453,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       });
     });
 
-    this.Editor.Tooltip.onHover(this.nodes.conversionToggler, I18n.ui(I18nInternalNS.ui.inlineToolbar.converter, 'Convert to'), {
+    this.tooltip.onHover(this.nodes.conversionToggler, I18n.ui(I18nInternalNS.ui.inlineToolbar.converter, 'Convert to'), {
       placement: 'top',
       hidingDelay: 100,
     });
@@ -508,13 +463,14 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    * Changes Conversion Dropdown content for current block's Tool
    */
   private setConversionTogglerContent(): void {
-    const { BlockManager, Tools } = this.Editor;
-    const toolName = BlockManager.currentBlock.name;
+    const { BlockManager } = this.Editor;
+    const { currentBlock } = BlockManager;
+    const toolName = currentBlock.name;
 
     /**
      * If tool does not provide 'export' rule, hide conversion dropdown
      */
-    const conversionConfig = Tools.available[toolName][Tools.INTERNAL_SETTINGS.CONVERSION_CONFIG] || {};
+    const conversionConfig = currentBlock.tool.conversionConfig;
     const exportRuleDefined = conversionConfig && conversionConfig.export;
 
     this.nodes.conversionToggler.hidden = !exportRuleDefined;
@@ -523,14 +479,10 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     /**
      * Get icon or title for dropdown
      */
-    const toolSettings = Tools.getToolSettings(toolName);
-    const toolboxSettings = Tools.available[toolName][Tools.INTERNAL_SETTINGS.TOOLBOX] || {};
-    const userToolboxSettings = toolSettings.toolbox || {};
+    const toolboxSettings = currentBlock.tool.toolbox || {};
 
     this.nodes.conversionTogglerContent.innerHTML =
-      userToolboxSettings.icon ||
       toolboxSettings.icon ||
-      userToolboxSettings.title ||
       toolboxSettings.title ||
       _.capitalize(toolName);
   }
@@ -563,21 +515,8 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     this.nodes.actions.innerHTML = '';
     this.toolsInstances = new Map();
 
-    /**
-     * Filter buttons if Block Tool pass config like inlineToolbar=['link']
-     * Else filter them according to the default inlineToolbar property.
-     *
-     * For this moment, inlineToolbarOrder could not be 'false'
-     * because this method will be called only if the Inline Toolbar is enabled
-     */
-    const inlineToolbarOrder = this.getInlineToolbarSettings(currentBlock.name) as string[];
-
-    inlineToolbarOrder.forEach((toolName) => {
-      const toolSettings = this.Editor.Tools.getToolSettings(toolName);
-      const tool = this.Editor.Tools.constructInline(this.Editor.Tools.inline[toolName], toolName, toolSettings);
-
-      this.addTool(toolName, tool);
-      tool.checkState(SelectionUtils.get());
+    Array.from(currentBlock.tool.inlineTools.values()).forEach(tool => {
+      this.addTool(tool);
     });
 
     /**
@@ -589,76 +528,39 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
   /**
    * Add tool button and activate clicks
    *
-   * @param {string} toolName - name of Tool to add
-   * @param {InlineTool} tool - Tool class instance
+   * @param {InlineTool} tool - InlineTool object
    */
-  private addTool(toolName: string, tool: InlineTool): void {
-    const {
-      Listeners,
-      Tools,
-      Tooltip,
-    } = this.Editor;
-
-    const button = tool.render();
+  private addTool(tool: InlineTool): void {
+    const instance = tool.create();
+    const button = instance.render();
 
     if (!button) {
-      _.log('Render method must return an instance of Node', 'warn', toolName);
+      _.log('Render method must return an instance of Node', 'warn', tool.name);
 
       return;
     }
 
-    button.dataset.tool = toolName;
+    button.dataset.tool = tool.name;
     this.nodes.buttons.appendChild(button);
-    this.toolsInstances.set(toolName, tool);
+    this.toolsInstances.set(tool.name, instance);
 
-    if (_.isFunction(tool.renderActions)) {
-      const actions = tool.renderActions();
+    if (_.isFunction(instance.renderActions)) {
+      const actions = instance.renderActions();
 
       this.nodes.actions.appendChild(actions);
     }
 
-    Listeners.on(button, 'click', (event) => {
-      this.toolClicked(tool);
+    this.listeners.on(button, 'click', (event) => {
+      this.toolClicked(instance);
       event.preventDefault();
     });
 
-    /**
-     * Enable shortcuts
-     * Ignore tool that doesn't have shortcut or empty string
-     */
-    const toolSettings = Tools.getToolSettings(toolName);
-
-    let shortcut = null;
-
-    /**
-     * Get internal inline tools
-     */
-    const internalTools: string[] = Object
-      .entries(Tools.internalTools)
-      .filter(([, toolClass]: [string, ToolConstructable | ToolSettings]) => {
-        if (_.isFunction(toolClass)) {
-          return toolClass[Tools.INTERNAL_SETTINGS.IS_INLINE];
-        }
-
-        return (toolClass as ToolSettings).class[Tools.INTERNAL_SETTINGS.IS_INLINE];
-      })
-      .map(([ name ]: [string, InlineToolConstructable | ToolSettings]) => name);
-
-    /**
-     * 1) For internal tools, check public getter 'shortcut'
-     * 2) For external tools, check tool's settings
-     * 3) If shortcut is not set in settings, check Tool's public property
-     */
-    if (internalTools.includes(toolName)) {
-      shortcut = this.inlineTools[toolName][Tools.INTERNAL_SETTINGS.SHORTCUT];
-    } else if (toolSettings && toolSettings[Tools.USER_SETTINGS.SHORTCUT]) {
-      shortcut = toolSettings[Tools.USER_SETTINGS.SHORTCUT];
-    } else if (tool.shortcut) {
-      shortcut = tool.shortcut;
-    }
+    const shortcut = this.getToolShortcut(tool.name);
 
     if (shortcut) {
-      this.enableShortcuts(tool, shortcut);
+      try {
+        this.enableShortcuts(instance, shortcut);
+      } catch (e) {}
     }
 
     /**
@@ -667,7 +569,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     const tooltipContent = $.make('div');
     const toolTitle = I18n.t(
       I18nInternalNS.toolNames,
-      Tools.toolsClasses[toolName][Tools.INTERNAL_SETTINGS.TITLE] || _.capitalize(toolName)
+      tool.title || _.capitalize(tool.name)
     );
 
     tooltipContent.appendChild($.text(toolTitle));
@@ -678,10 +580,40 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       }));
     }
 
-    Tooltip.onHover(button, tooltipContent, {
+    this.tooltip.onHover(button, tooltipContent, {
       placement: 'top',
       hidingDelay: 100,
     });
+
+    instance.checkState(SelectionUtils.get());
+  }
+
+  /**
+   * Get shortcut name for tool
+   *
+   * @param toolName — Tool name
+   */
+  private getToolShortcut(toolName): string | void {
+    const { Tools } = this.Editor;
+
+    /**
+     * Enable shortcuts
+     * Ignore tool that doesn't have shortcut or empty string
+     */
+    const tool = Tools.inlineTools.get(toolName);
+
+    /**
+     * 1) For internal tools, check public getter 'shortcut'
+     * 2) For external tools, check tool's settings
+     * 3) If shortcut is not set in settings, check Tool's public property
+     */
+    const internalTools = Tools.internal.inlineTools;
+
+    if (Array.from(internalTools.keys()).includes(toolName)) {
+      return this.inlineTools[toolName][CommonInternalSettings.Shortcut];
+    }
+
+    return tool.shortcut;
   }
 
   /**
@@ -690,8 +622,8 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    * @param {InlineTool} tool - Tool instance
    * @param {string} shortcut - shortcut according to the ShortcutData Module format
    */
-  private enableShortcuts(tool: InlineTool, shortcut: string): void {
-    this.Editor.Shortcuts.add({
+  private enableShortcuts(tool: IInlineTool, shortcut: string): void {
+    Shortcuts.add({
       name: shortcut,
       handler: (event) => {
         const { currentBlock } = this.Editor.BlockManager;
@@ -710,15 +642,14 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
          */
         // if (SelectionUtils.isCollapsed) return;
 
-        const toolSettings = this.Editor.Tools.getToolSettings(currentBlock.name);
-
-        if (!toolSettings || !toolSettings[this.Editor.Tools.USER_SETTINGS.ENABLED_INLINE_TOOLS]) {
+        if (!currentBlock.tool.enabledInlineTools) {
           return;
         }
 
         event.preventDefault();
         this.toolClicked(tool);
       },
+      on: this.Editor.UI.nodes.redactor,
     });
   }
 
@@ -727,7 +658,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    *
    * @param {InlineTool} tool - Tool's instance
    */
-  private toolClicked(tool: InlineTool): void {
+  private toolClicked(tool: IInlineTool): void {
     const range = SelectionUtils.range;
 
     tool.surround(range);
@@ -747,16 +678,14 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    * Get inline tools tools
    * Tools that has isInline is true
    */
-  private get inlineTools(): { [name: string]: InlineTool } {
+  private get inlineTools(): { [name: string]: IInlineTool } {
     const result = {};
 
-    for (const tool in this.Editor.Tools.inline) {
-      if (Object.prototype.hasOwnProperty.call(this.Editor.Tools.inline, tool)) {
-        const toolSettings = this.Editor.Tools.getToolSettings(tool);
-
-        result[tool] = this.Editor.Tools.constructInline(this.Editor.Tools.inline[tool], tool, toolSettings);
-      }
-    }
+    Array
+      .from(this.Editor.Tools.inlineTools.entries())
+      .forEach(([name, tool]) => {
+        result[name] = tool.create();
+      });
 
     return result;
   }