+++ src/components/selection.ts
@@ -1,5 +1,5 @@
 /**
- * TextRange interface fot IE9-
+ * TextRange interface for IE9-
  */
 import * as _ from './utils';
 import $ from './dom';
@@ -125,7 +125,7 @@ export default class SelectionUtils {
 
     let editorZone = null;
 
-    if (selectedNode) {
+    if (selectedNode && selectedNode instanceof Element) {
       editorZone = selectedNode.closest(`.${SelectionUtils.CSS.editorZone}`);
     }